optimize

+++ include/queue_spsc.h
@@ -122,7 +122,7 @@ template <typename T>
 class qring {
 public:
     using ei_t = std::uint8_t;
-    using ti_t = std::uint16_t;
+    using ti_t = std::uint32_t;
 
     enum : std::size_t {
         elem_max = (std::numeric_limits<ei_t>::max)() + 1, // 255 + 1
